refactor: split activate() from handlers (#11858)
* refactor: split activate() from handlers

* test: fix ut

* test: add ut

* fix: fix ut
