+++ packages/vscode-extension/src/extension.ts
@@ -74,6 +74,7 @@ import {
   workspaceUri,
 } from "./globalVariables";
 import * as handlers from "./handlers";
+import { activate as activateHandlers } from "./handlers/activate";
 import { checkCopilotAccessHandler } from "./handlers/checkCopilotAccess";
 import { checkCopilotCallback } from "./handlers/checkCopilotCallback";
 import { checkSideloadingCallback } from "./handlers/checkSideloading";
@@ -161,7 +162,7 @@ export async function activate(context: vscode.ExtensionContext) {
   }
 
   // Call activate function of toolkit core.
-  handlers.activate();
+  activateHandlers();
 
   // Init VSC context key
   await initializeContextKey(context, isTeamsFxProject);
